requirements.txt
cmake==3.21.1.post1
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Jinja2==3.0.1
numpy~=1.19.5
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
Pillow==8.3.1
PyMySQL==1.0.2
SQLAlchemy~=1.4.23
Werkzeug~=2.0.1
requests~=2.26.0
Keras~=2.4.3